move _allocate_section_extents from both main widgets to a new _layout_utils file
