+++ src/napari_metadata/viewer_widgets/_viewer_metadata.py
@@ -13,6 +13,7 @@ from qtpy.QtWidgets import (
     QWidget,
 )
 
+from napari_metadata._layout_utils import _allocate_section_extents
 from napari_metadata.viewer_widgets._dims import DimsWidget
 from napari_metadata.viewer_widgets._scale_bar import ScaleBarWidget
 from napari_metadata.widgets._containers import (
@@ -25,62 +26,6 @@ if TYPE_CHECKING:
     from napari.components import ViewerModel
 
 _SECTIONS_SPACING = 3
-
-
-def _allocate_section_extents(
-    *,
-    expanded: list[bool],
-    collapsed_extents: list[int],
-    preferred_extents: list[int],
-    available: int,
-    spacing: int,
-) -> list[int]:
-    """Distribute available pixels across collapsed and expanded sections."""
-    extents = collapsed_extents.copy()
-    expanded_indices = [
-        index for index, is_expanded in enumerate(expanded) if is_expanded
-    ]
-    if not expanded_indices:
-        return extents
-
-    collapsed_total = sum(
-        extent
-        for extent, is_expanded in zip(
-            collapsed_extents, expanded, strict=True
-        )
-        if not is_expanded
-    )
-    usable = max(available - spacing - collapsed_total, 0)
-
-    preferred_by_index = {
-        index: max(preferred_extents[index], collapsed_extents[index])
-        for index in expanded_indices
-    }
-    minimum_total = sum(collapsed_extents[index] for index in expanded_indices)
-    if usable <= minimum_total:
-        return extents
-
-    preferred_total = sum(
-        preferred_by_index[index] for index in expanded_indices
-    )
-    if usable >= preferred_total:
-        for index in expanded_indices:
-            extents[index] = preferred_by_index[index]
-        return extents
-
-    remaining = usable
-    for offset, index in enumerate(
-        sorted(expanded_indices, key=lambda item: preferred_by_index[item])
-    ):
-        share = remaining // (len(expanded_indices) - offset)
-        extent = max(
-            collapsed_extents[index],
-            min(preferred_by_index[index], share),
-        )
-        extents[index] = extent
-        remaining -= extent
-
-    return extents
 
 
 class ViewerMetadataWidget(QWidget):
